services/note-waver/requirements.txt
fastapi~=0.114.2
uvicorn==0.22.0
python-dotenv~=1.0.1
requests~=2.25.1
valkey~=6.0.2